Test/requirements.txt
flask==2.3.3
torch>=1.9.0
torchvision>=0.10.0
transformers>=4.21.0
pillow>=9.0.0
numpy>=1.21.0
opencv-python>=4.5.0
paddlepaddle>=2.4.0
paddleocr>=2.6.0
groundingdino-py>=0.4.0
supervision>=0.16.0
ultralytics>=8.0.0
segment-anything>=1.0
sam2>=0.1.0